# memmetrics: simplify locking and solve data race